src/backend/requirements.txt
# Core
fastapi[standard]==0.121.0
asyncpg==0.30.0
SQLAlchemy==2.0.44
alembic==1.17.1

# Pydantic Settings 
pydantic-settings==2.11.0

# Loguru
loguru==0.7.3

# Tests
httpx==0.28.1,
pytest==8.4.2,
pytest-asyncio==1.2.0

# Websockets
websockets==15.0.1

# For importing images to the printer
pillow==12.0.0